EPS-20: move CourseWithStudentExtractor to another package
